demos/agno-workflows/pyproject.toml
[project]
name = "agno-workflows"
version = "0.1.0"
requires-python = ">=3.13,<3.14"
dependencies = [
    "agno[os] (>=2.5.10,<3.0.0)",
    "anthropic (>=0.84.0,<1.0.0)",
    "psycopg[binary] (>=3.1.0,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
]

[tool.poetry]
packages = []

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
